server/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
sqlmodel>=0.0.14
pyyaml>=6.0
aiofiles>=23.2
requests>=2.31
Pillow>=10.0
python-multipart>=0.0.6
